refactor: 将 CreatorAudioDetailVO 移动到 vo 文件夹
Close #99
