+++ src/main/java/com/github/listen_to_me/service/IAudioInfoService.java
@@ -9,7 +9,6 @@ import com.baomidou.mybatisplus.extension.service.IService;
 import com.github.listen_to_me.domain.dto.AudioAuditDTO;
 import com.github.listen_to_me.domain.dto.AudioDTO;
 import com.github.listen_to_me.domain.dto.AudioUpdateDTO;
-import com.github.listen_to_me.domain.dto.CreatorAudioDetailVO;
 import com.github.listen_to_me.domain.entity.AudioInfo;
 import com.github.listen_to_me.domain.query.AudioSearchQuery;
 import com.github.listen_to_me.domain.query.AuditQuery;
@@ -20,6 +19,7 @@ import com.github.listen_to_me.domain.vo.AudioPublishVO;
 import com.github.listen_to_me.domain.vo.AudioStatusVO;
 import com.github.listen_to_me.domain.vo.AudioVO;
 import com.github.listen_to_me.domain.vo.AuditAudioVO;
+import com.github.listen_to_me.domain.vo.CreatorAudioDetailVO;
 import com.github.listen_to_me.domain.vo.CreatorAudioVO;
 
 /**
