+++ src/main/java/com/github/listen_to_me/service/impl/AudioInfoServiceImpl.java
@@ -28,7 +28,6 @@ import com.github.listen_to_me.common.util.RedisUtils;
 import com.github.listen_to_me.domain.dto.AudioAuditDTO;
 import com.github.listen_to_me.domain.dto.AudioDTO;
 import com.github.listen_to_me.domain.dto.AudioUpdateDTO;
-import com.github.listen_to_me.domain.dto.CreatorAudioDetailVO;
 import com.github.listen_to_me.domain.entity.AudioFolderRelation;
 import com.github.listen_to_me.domain.entity.AudioInfo;
 import com.github.listen_to_me.domain.entity.AudioLike;
@@ -46,6 +45,7 @@ import com.github.listen_to_me.domain.vo.AudioPublishVO;
 import com.github.listen_to_me.domain.vo.AudioStatusVO;
 import com.github.listen_to_me.domain.vo.AudioVO;
 import com.github.listen_to_me.domain.vo.AuditAudioVO;
+import com.github.listen_to_me.domain.vo.CreatorAudioDetailVO;
 import com.github.listen_to_me.domain.vo.CreatorAudioVO;
 import com.github.listen_to_me.mapper.AudioFolderRelationMapper;
 import com.github.listen_to_me.mapper.AudioInfoMapper;
